Moved src/model/geometrical_model classes to src/epipeODE/init in anticipation of the class rework
